Introduce errors module
